requirements.txt
acres==0.5.0
aiofiles==25.1.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
filelock==3.20.0
frontend==0.0.3
h11==0.16.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
looseversion==1.3.0
lxml==6.0.2
networkx==3.6
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.0.0
prov==2.1.1
puremagic==1.30
pydot==4.0.1
PyMuPDF==1.26.6
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.4
rdflib==7.4.0
requests==2.32.5
scipy==1.16.3
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tools==1.0.13
traits==7.0.2
typing==3.7.4.3
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
